rename: with access token 파일 위치 변경
